Rename file - fix tests
